requirements.txt
requests
python-dotenv
weaviate-client>=4.6,<5
beautifulsoup4
lxml[html_clean]
newspaper3k
openai